requirements.txt
flask==2.3.3
openai-whisper==20231117
pydub==0.25.1
ffmpeg-python==0.2.0
Werkzeug==2.3.7
gunicorn==20.1.0
python-dotenv==1.0.0 